Bugfix: Hands not recognized outside the screen viewport.
Fix: Added more space in the sides (left and right) to let the camera and the canvas work even if the hand is on
the border. Added a constant to set the extra width to 600px, which is good for a hand size close to the screen.
Camera capturing area and canvas drawing area increased in the X axis.